+++ mpich2/test/mpi/io/large_count.c
@@ -67,7 +67,7 @@ int main(int argc, char **argv)
         if (rank == 0) {
             printf("nprocs=%d NVARS=%d LEN=%d\n", nprocs, NVARS, LEN);
             printf("Expecting file size=%lld bytes (%.1f MB, %.1f GB)\n",
-                   fsize, (float) fsize / 1048576, (float) fsize / 1073741824);
+                   (long long) fsize, (float) fsize / 1048576, (float) fsize / 1073741824);
             printf("Each global variable is of size %d bytes (%.1f MB)\n",
                    gsize[0] * gsize[1], (float) gsize[0] * gsize[1] / 1048576);
             printf("Each process writes %zd bytes (%.1f MB, %.1f GB)\n",
@@ -150,7 +150,7 @@ int main(int argc, char **argv)
     /* do it again but use the non-blocking collectives */
     err = MPI_File_seek(fh, 0, MPI_SEEK_SET);
     ERROR("MPI_File_seek");
-    MPI_Request  req;
+    MPI_Request req;
     err = MPI_File_iwrite_all(fh, buf, 1, buftype, &req);
     ERROR("MPI_File_write_all");
     MPI_Wait(&req, &status);
